+++ lie_nn/__init__.py
@@ -9,7 +9,8 @@ from ._src.rep import (
     SumRep,
     QRep,
 )
-from ._src.change_basis import change_basis, project
+from ._src.change_basis import change_basis
+from ._src.project import project
 from ._src.reduce import reduce
 from ._src.change_algebra import change_algebra
 from ._src.tensor_product import tensor_product, tensor_power
